Cleaned code a bit
